Renamed `helpers` to `connections`
