+++ bxserum/connections/ws.go
@@ -1,4 +1,4 @@
-package helpers
+package connections
 
 import (
 	"context"
