--- bxserum/helpers/grpc.go
@@ -1,1 +1,0 @@
-package helpers
